Abstract template helper
It seems insane to include the entire application just to process an ERB
template. It makes far more sense to create a helper to contain the
code needed by the template and bind it.
